Update README.md
The primary objective of this project is, in essence, to offer users an immersive and captivating storytelling experience that defies conventional boundaries. 
By wedding the abilities of Transformers , GPT-2, and web advances, this try looks to reclassify the scene of continuous story age, representing the agreeable mix of computerized reasoning and client.
The integration not only emphasizes the symbiotic relationship between artificial intelligence and web development, but it also demonstrates the technical prowess of AI-driven content creation.
The core objective lies in demonstrating the fusion of AI driven content creation and web development. The user-friendly interface, developed using HTML, CSS, and JavaScript, empowers users to input prompts and receive generated stories dynamically.